New redis data model for jobids
Jobid-keys now have hash values, and we use hset and hgetall to manage
this.

REDIS DATA MODEL
----------------
Jobids are UUID strings. These jobids are used as redis keys.
The values for jobid keys are redis hashes. The following fields should ALWAYS
be present:

1) type (ASR, ASR_SEGMENTS etc.)
2) status (done, pending etc.)
3) processing_started (unix timestamp)

Additionally, if the following are known, they will be in the hash (rather than in
a json blob somewhere)

4) processing_finished
5) segments (json list of {duration, jobid})
6) response (json object)